@@ -28,7 +28,7 @@
     "class-validator": "^0.14.0",
     "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
-    "eslint": "^8.23.0",
+    "eslint": "^8.44.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-config-standard-with-typescript": "^22.0.0",
     "eslint-plugin-import": "^2.26.0",
@@ -47,12 +47,12 @@
     "validator": "^13.9.0"
   },
   "dependencies": {
+    "@faker-js/faker": "^8.0.2",
     "@types/multer": "^1.4.7",
     "axios": "^1.6.0",
     "graphql": "^15.8.0",
     "helmet": "^7.0.0",
     "multer": "^1.4.5-lts.1",
-    "mysql": "^2.18.1",
     "typeorm-extension": "^3.0.1"
   }
 }
