@@ -1,6 +1,6 @@
 import express from 'express';
 import dataSource from './dataSource';
-const path = require('path');
+import path from 'path';
 import cookieParser from 'cookie-parser';
 import dotenv from 'dotenv';
 import {
@@ -10,12 +10,14 @@ import {
    citiesRoutes,
    categoriesRoutes,
 } from './routes';
+//import { seed } from './seed';
 import helmet from 'helmet';
 
 dotenv.config();
 
 const app = express();
 app.use(express.json());
+
 app.use(cookieParser());
 app.use(
    helmet({
@@ -23,7 +25,17 @@ app.use(
    })
 );
 app.use((req, res, next) => {
-   res.setHeader('Access-Control-Allow-Origin', 'http://localhost:3000');
+   const corsWhitelist = [
+      'http://localhost:3000',
+      'https://lamarr4.wns.wilders.dev',
+   ];
+
+   if (
+      req.headers.origin !== undefined &&
+      corsWhitelist.includes(req.headers.origin)
+   ) {
+      res.header('Access-Control-Allow-Origin', req.headers.origin);
+   }
    res.setHeader('Access-Control-Allow-Credentials', 'true');
    res.setHeader(
       'Access-Control-Allow-Headers',
@@ -47,6 +59,15 @@ const start = async (): Promise<void> => {
    const port = 5000;
 
    await dataSource.initialize();
+
+   /*  if (process.env.NODE_ENV !== 'production') {
+      try {
+         await seed();
+      } catch (error) {
+         console.log('Seed error: ' + error);
+      }
+   } */
+
    app.listen({ port }, () => {
       console.log(`Backend app ready at http://localhost:${port}`);
    });
