@@ -26,7 +26,7 @@ export const PoiFactory = setSeederFactory(Poi, (faker: Faker) => {
   };
   poi.description = faker.lorem.sentence();
   poi.image = faker.image.urlPicsumPhotos();
-  poi.is_accepted = faker.datatype.boolean();
+  poi.isAccepted = faker.datatype.boolean();
 
   return poi;
 });
