@@ -1,10 +1,17 @@
+/* eslint-disable no-lone-blocks */
 // import React, { useContext } from 'react'
-import Header from "../../components/common/header/Header";
-import { UsersContext, UserProvider } from "../../contexts/UserContext";
-import { useNavigate } from "react-router-dom";
+
+import InteractiveMap from "../../components/interactiveMap/InteractiveMap";
 import { useContext, useEffect, useState } from "react";
-import Footer from "../../components/common/footer/Footer";
-import "./homePage.scss";
+
+import SearchPOI from "../SearchPOI/SearchPOI";
+
+import styles from "./homePage.module.scss";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+
+import { faPaperPlane } from "@fortawesome/free-solid-svg-icons";
+import { useNavigate } from "react-router-dom";
+import { UsersContext } from "../../contexts/UserContext";
 
 const HomePage = () => {
   const [windowSize, setWindowSize] = useState<number>(window.innerWidth);
@@ -23,29 +30,53 @@ const HomePage = () => {
   console.log(isAuthenticated);
 
   return (
-    <UserProvider>
-      <Header size={windowSize > 768 ? "desktop" : "mobile"} />
-      <div className="container-link-map">
-        {/* <InteractiveMap /> */}
-        <button>
-          <span>Voir la carte</span>
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            width="13"
-            height="17"
-            fill="none"
-            viewBox="0 0 13 17"
+    <>
+      {windowSize > 768 ? (
+        <>
+          <section className={`${styles.backgroundWave}`}>
+            <div className={`presentationText ${styles.presentationSection} `}>
+              <p>
+                Lorem ipsum dolor sit amet, consectetur adipisicing elit. Illum
+                blanditiis aperiam ad, aspernatur incidunt voluptatem ea
+                molestias nemo rem ratione nesciunt neque corporis nihil
+                recusandae cum minima. Id, voluptatum tenetur?
+              </p>
+              <button
+                className={`buttonHomePage textButton ${styles.presentationButton}`}
+              >
+                <FontAwesomeIcon icon={faPaperPlane} />
+                Explorez
+              </button>
+            </div>
+          </section>
+          <section>
+            <div className={`${styles.mapSection}`}>
+              <InteractiveMap />
+            </div>
+            <div className={`${styles.pictoPlane}`}>
+              <img src="/picto_avion.png" alt="picto avion" />
+            </div>
+          </section>
+          <section className={`${styles.searchSection}`}>
+            <SearchPOI />
+          </section>
+        </>
+      ) : (
+        <section className={`${styles.mobileSection}`}>
+          <h1>CITY GUIDE</h1>
+          <img src="/wave_with_map_mobile.svg" alt="wave with map" />
+          <button
+            className={`buttonHomePage textButton ${styles.presentationButton}`}
           >
-            <path
-              fill="#090F43"
-              d="M12.477 6.52c0 3.445-6.238 9.972-6.238 9.972S0 9.965 0 6.52a6.239 6.239 0 1 1 12.477 0Z"
-            />
-            <circle cx="6.238" cy="6.064" r="2.322" fill="#F3DF24" />
-          </svg>
-        </button>
-      </div>
-      <Footer />
-    </UserProvider>
+            Explorez
+          </button>
+          <h2 className="subtitleCategoryDark">
+            Trouvez votre point d'intérêt !
+          </h2>
+          <SearchPOI />
+        </section>
+      )}
+    </>
   );
 };
 
